refactor: extract hydration logic to a new module
